+++ R/shift.R
@@ -1,7 +1,7 @@
 #' Shift values in a single-case data file
 #'
-#' *This function has been replaced by the much more versatile `transform`
-#' function.*
+#' *This function has been superseded by the much more versatile 
+#' \code{\link{transform.scdf}} function.*
 #' Shifting the values might be helpful in cases where the measurement time
 #' is given as a time variable (see example below).
 #'
@@ -20,8 +20,9 @@
 #' plm(ex)
 #' 
 #' # Please use transform instead:
-#' ex <- transform(example_A24, year = year - 1996)
-#' plm(ex)
+#' example_A24 %>%
+#'   transform(year = year - 1996) %>%
+#'   plm()
 shift <- function(data, value, var) {
   .deprecated_warning("transform", "shift")
   if (missing(var)) var <- scdf_attr(data, .opt$mt)